requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.5
python-dotenv>=1.0.0
requests
logging
jishaku>=2.5.1
humanize>=4.7.0
pytz>=2023.3
colorlog>=6.7.0
asyncio>=3.4.3
flask~=3.1.0
yt-dlp~=2024.12.13
spotipy~=2.25.0
discord~=2.3.2
google-api-python-client~=2.169.0
psutil
supabase>=2.3.4
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0  # For JWT support